Push dead message handlers to the package level
